pyproject.toml
[build-system]
requires = ["setuptools>=80", "setuptools-scm>=8", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "CO-UCAgent"
description = "Context Optimized UCAgent - enhanced context management for verification"
authors = [
    {name = "CO-UCAgent Contributors"},
]
license = {text = "Apache-2.0"}
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
dependencies = [
    "langchain>=1.2.0",
    "langchain_community>=0.3.31",
    "openai>=1.109.1",
    "langgraph>=1.0.5",
    "langchain_openai>=1.1.6",
    "langmem>=0.0.30",
    "urwid>=3.0.3",
    "mem0ai>=0.1.118",
    "langchain_mcp_adapters>=0.1.11",
    "psutil>=7.1.0",
    "langfuse>=3.9.1",
    "gitpython>=3.1.43",
    "pytoffee@git+https://github.com/XS-MLVP/toffee@master",
    "toffee-test@git+https://github.com/XS-MLVP/toffee-test@master",
]
dynamic = ["version"]

[project.scripts]
ucagent = "ucagent.cli:main"

[project.urls]
Homepage = "https://github.com/XS-MLVP/UCAgent"

[tool.setuptools_scm]
fallback_version = "0.9.1"
version_file = "ucagent/_version.py"
version_scheme = "guess-next-dev"
local_scheme = "node-and-date"

[tool.setuptools]
packages = ["ucagent"]
